fix repeated code and PEP8 stuff
