+++ yacht/download_pretrained_ref_db.py
@@ -5,6 +5,7 @@ from loguru import logger
 import sys
 import os
 import zipfile
+from .utils import create_output_folder, check_download_args
 
 # Configure Loguru logger
 logger.remove()
@@ -21,6 +22,7 @@ def add_arguments(parser):
     parser.add_argument("--ani_thresh", type=float, choices=[0.80, 0.95, 0.995, 0.9995], required=True)
     parser.add_argument("--outfolder", help="Output folder for downloaded files.", default=".")
 
+
 def fetch_zenodo_records():
     logger.info("Fetching list of files from Zenodo community 'yacht'")
     try:
@@ -31,19 +33,21 @@ def fetch_zenodo_records():
         logger.error(f"Error fetching data from Zenodo: {e}")
         return []
 
+
 def generate_download_url(args):
     if args.database == "genbank":
         if args.db_version == "genbank-2022.03":
             return f"{args.db_version}-{args.ncbi_organism}-k{args.k}"
         else:
-            logger.error(f"Invalid GenBank version: {args.db_version}. Now only support genbank-2022.03.")
+            logger.error(f"Invalid GenBank version: {args.db_version}. We currently only support genbank-2022.03.")
             return None
     else:
         if args.db_version == "rs214":
             return f"{args.database}-{args.db_version}-reps.k{args.k}"
         else:
-            logger.error(f"Invalid GTDB version: {args.db_version}. Now only support rs214.")
+            logger.error(f"Invalid GTDB version: {args.db_version}. We currently only support rs214.")
             return None
+
 
 def download_file(url, output_path):
     if os.path.exists(output_path):
@@ -61,10 +65,6 @@ def download_file(url, output_path):
         logger.error(f"Failed to download {url}: {e}")
         return False
 
-def create_output_folder(outfolder):
-    if not os.path.exists(outfolder):
-        logger.info(f"Creating output folder: {outfolder}")
-        os.makedirs(outfolder)
 
 def unzip_file(file_path, extract_to):
     try:
@@ -74,50 +74,27 @@ def unzip_file(file_path, extract_to):
     except zipfile.BadZipFile:
         logger.error(f"Failed to unzip {file_path}. It might not be a zip file.")
 
-def create_output_folder(outfolder):
-    if not os.path.exists(outfolder):
-        logger.info(f"Creating output folder: {outfolder}")
-        os.makedirs(outfolder)
 
 def main(args):
-    ## Check if the input arguments are valid
-    if args.database not in ["genbank", "gtdb"]:
-        logger.error(f"Invalid database: {args.database}. Now only support genbank and gtdb.")
-        os.exit(1)
+    # Check if the input arguments are valid
+    check_download_args(args)
 
-    if args.k not in [21, 31, 51]:
-        logger.error(f"Invalid k: {args.k}. Now only support 21, 31, and 51.")
-        os.exit(1)
-
-    if args.database == "genbank":
-        if args.ncbi_organism is None:
-            logger.warning("No NCBI organism specified using parameter --ncbi_organism. Use default: bacteria")
-            args.ncbi_organism = "bacteria"
-            
-        if args.ncbi_organism not in ["archaea", "bacteria", "fungi", "virus", "protozoa"]:            
-            logger.error(f"Invalid NCBI organism: {args.ncbi_organism}. Now only support archaea, bacteria, fungi, virus, and protozoa.")
-            os.exit(1)
-        
-        if args.ncbi_organism == "virus":
-            logger.error("We now have support for virus database.")
-            os.exit(1)
-
-    ## Generate download URL
+    # Generate download URL
     file_prefix = generate_download_url(args)
     if not file_prefix:
         os.exit(1)
 
-    ## Fetch list of files from Zenodo
+    # Fetch list of files from Zenodo
     zenodo_records = fetch_zenodo_records()
     if not zenodo_records:
         logger.error("No records fetched from Zenodo. Exiting.")
         os.exit(1)
     current_pretrained_db_list = [record['title'] for record in zenodo_records]
 
-    ## Create output folder if not exists
+    # Create output folder if not exists
     create_output_folder(args.outfolder)
 
-    ## Check if the specified db exists in Zenodo
+    # Check if the specified db exists in Zenodo
     available_files = [filename for filename in current_pretrained_db_list if file_prefix in filename]
     if len(available_files) == 0:
         logger.error(f"No pretrained database with prefix {file_prefix} found on Zenodo.")
@@ -127,18 +104,19 @@ def main(args):
         available_files_with_ani = [filename for filename in available_files if f"{args.ani_thresh}" in filename]
         available_ani_thresh = [float(x.split('_')[-2]) for x in available_files]
         if len(available_files_with_ani) == 0:
-            logger.error(f"No pretrained database found for {file_prefix}_{args.ani_thresh}_pretrained.zip on Zenodo. Now only support {available_ani_thresh} for {file_prefix}.")
+            logger.error(f"No pretrained database found for {file_prefix}_{args.ani_thresh}_pretrained.zip on Zenodo. "
+                         f"We currenlty only support {available_ani_thresh} for {file_prefix}.")
             os.exit(1)
         else:
             file_name_to_search = available_files_with_ani[0]    
 
-    ## Check if the file already exists
+    # Check if the file already exists
     output_path = os.path.join(args.outfolder, file_name_to_search)
     if os.path.exists(output_path):
         logger.info(f"File {output_path} already exists. Skipping download.")
         return
 
-    ## Download the file
+    # Download the file
     file_url = next((file.get('links', {}).get('self') for record in zenodo_records for file in record.get('files', [])
                      if file_name_to_search in file.get('key', '')), None)
 
@@ -146,6 +124,7 @@ def main(args):
         unzip_file(output_path, args.outfolder)
     else:
         logger.warning(f"File '{file_name_to_search}' not found in Zenodo records.")
+
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
