+++ yacht/utils.py
@@ -345,3 +345,38 @@ def get_cami_profile(cami_content: List[str]) -> List[Tuple[str, Dict[str, str],
         raise RuntimeError
 
     return samples_list
+
+
+def create_output_folder(outfolder):
+    """
+    Helper function that creates the output folder if it does not exist.
+    :param outfolder: location of output folder
+    :return: None
+    """
+    if not os.path.exists(outfolder):
+        logger.info(f"Creating output folder: {outfolder}")
+        os.makedirs(outfolder)
+
+
+def check_download_args(args):
+    if args.database not in ["genbank", "gtdb"]:
+        logger.error(f"Invalid database: {args.database}. Now only support genbank and gtdb.")
+        os.exit(1)
+
+    if args.k not in [21, 31, 51]:
+        logger.error(f"Invalid k: {args.k}. Now only support 21, 31, and 51.")
+        os.exit(1)
+
+    if args.database == "genbank":
+        if args.ncbi_organism is None:
+            logger.warning("No NCBI organism specified using parameter --ncbi_organism. Using the default: bacteria")
+            args.ncbi_organism = "bacteria"
+
+        if args.ncbi_organism not in ["archaea", "bacteria", "fungi", "virus", "protozoa"]:
+            logger.error(
+                f"Invalid NCBI organism: {args.ncbi_organism}. Now only support archaea, bacteria, fungi, virus, and protozoa.")
+            os.exit(1)
+
+        if args.ncbi_organism == "virus":
+            logger.error("We now have support for virus database.")
+            os.exit(1)
